+++ internal/elasticsearch/security/user.go
@@ -116,17 +116,13 @@ func resourceSecurityUserPut(ctx context.Context, d *schema.ResourceData, meta i
 	var user models.User
 	user.Username = usernameId
 
-	if v, ok := d.GetOk("password"); ok {
+	if v, ok := d.GetOk("password"); ok && d.HasChange("password") {
 		password := v.(string)
-		if d.HasChange("password") {
-			user.Password = &password
-		}
+		user.Password = &password
 	}
-	if v, ok := d.GetOk("password_hash"); ok {
+	if v, ok := d.GetOk("password_hash"); ok && d.HasChange("password_hash") {
 		pass_hash := v.(string)
-		if d.HasChange("password_hash") {
-			user.PasswordHash = &pass_hash
-		}
+		user.PasswordHash = &pass_hash
 	}
 
 	if v, ok := d.GetOk("email"); ok {