+++ examples/resources/elasticstack_elasticsearch_security_role/import.sh
@@ -1,1 +1,1 @@
-terraform import elasticstack_elasticsearch_security_role <cluster_uuid>/<role name>
+terraform import elasticstack_elasticsearch_security_role.my_role <cluster_uuid>/<role name>